+++ tanks-game/_source/generalized-lightning.h
@@ -36,7 +36,7 @@ protected:
 	double stateMultiplier[2]; //length = 2 because bool bolt action
 
 	unsigned int maxBolts; //this is maximum amount of normal bolts; the lightning can make any number of bolts when it has to destroy a bullet or tank
-	double lengthOfBolt;
+	float lengthOfBolt;
 	std::vector<LightningBolt*> bolts;
 	virtual void clearBolts();
 	double boltTick;
@@ -44,7 +44,7 @@ protected:
 	bool boltsNeeded; //= false; //if the lightning hits something, this is changed, and no random bolts will be made; resets every boltCycle ticks
 
 	virtual void refreshBolt(LightningBolt*) const = 0; //"redraw" a bolt //this is the hardest thing to generalize, so... copy and paste
-	virtual int getDefaultNumBoltPoints(double horzDist) const; //number of points that make up a bolt
+	virtual int getDefaultNumBoltPoints(float horzDist) const; //number of points that make up a bolt
 	virtual void pushBolt(LightningBolt*) = 0;
 	virtual void pushDefaultBolt(int num, bool randomize) = 0; //randomize should be true all of the time
 	std::vector<Game_ID> targetedObjects;
